+++ data/core/_main.cfg
@@ -8,7 +8,6 @@
 wesnoth.dofile 'lua/wml-tags.lua'
 wesnoth.dofile 'lua/feeding.lua'
 wesnoth.dofile 'lua/diversion.lua'
-wesnoth.dofile 'lua/backwards-compatibility.lua'
 >>
 [/lua]
