+++ src/scripting/lua_kernel_base.cpp
@@ -39,7 +39,6 @@
 #include "scripting/lua_map_location_ops.hpp"
 #include "scripting/lua_rng.hpp"
 #include "scripting/lua_widget.hpp"
-#include "scripting/lua_widget_methods.hpp" //intf_show_dialog
 #include "scripting/push_check.hpp"
 
 #include "game_version.hpp"                  // for do_version_check, etc
@@ -383,7 +382,7 @@ lua_kernel_base::lua_kernel_base()
 
 	cmd_log_ << "Initializing " << my_name() << "...\n";
 
-	// Define the CPP_function metatable ( so we can override print to point to a C++ member function, add "show_dialog" for this kernel, etc. )
+	// Define the CPP_function metatable ( so we can override print to point to a C++ member function, add certain functions for this kernel, etc. )
 	// Do it first of all in case C++ functions are ever used in the core Wesnoth libs loaded in the next step
 	cmd_log_ << "Adding boost function proxy...\n";
 
@@ -450,21 +449,9 @@ lua_kernel_base::lua_kernel_base()
 		{ "read_file",                &lua_fileops::intf_read_file          },
 		{ "canonical_path",           &lua_fileops::intf_canonical_path     },
 		{ "textdomain",               &lua_common::intf_textdomain   		},
-		//{ "get_dialog_value",         &lua_gui2::intf_get_dialog_value		},
-		//{ "set_dialog_tooltip",       &lua_gui2::intf_set_dialog_tooltip	},
-		//{ "set_dialog_active",        &lua_gui2::intf_set_dialog_active		},
-		//{ "set_dialog_visible",       &lua_gui2::intf_set_dialog_visible    },
-		//{ "add_dialog_tree_node",     &lua_gui2::intf_add_dialog_tree_node	},
-		//{ "set_dialog_callback",      &lua_gui2::intf_set_dialog_callback	},
-		//{ "set_dialog_canvas",        &lua_gui2::intf_set_dialog_canvas		},
-		//{ "set_dialog_focus",         &lua_gui2::intf_set_dialog_focus      },
-		//{ "set_dialog_markup",        &lua_gui2::intf_set_dialog_markup		},
-		//{ "set_dialog_value",         &lua_gui2::intf_set_dialog_value		},
-		//{ "remove_dialog_item",       &lua_gui2::intf_remove_dialog_item    },
 		{ "dofile",                   &dispatch<&lua_kernel_base::intf_dofile>           },
 		{ "require",                  &dispatch<&lua_kernel_base::intf_require>          },
 		{ "kernel_type",              &dispatch<&lua_kernel_base::intf_kernel_type>          },
-		{ "show_dialog",              &intf_show_dialog   },
 		{ "compile_formula",          &lua_formula_bridge::intf_compile_formula},
 		{ "eval_formula",             &lua_formula_bridge::intf_eval_formula},
 		{ "name_generator",           &intf_name_generator           },
